requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
ftfy==6.3.1
h11==0.16.0
hdbscan==0.8.41
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
open_clip_torch @ git+https://github.com/mlfoundations/open_clip.git@d3cdb734a2710feeb4c6307df037afa5f786a3e1
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
